don't attempt to reuse cached nonvalidated records
at least for now (for queries without +cd).
It wasn't complete, and it turned out to need more changes,
and the benefits would be rather limited.

+++ lib/layer/iterate.c
@@ -305,7 +305,7 @@ static inline uint8_t get_initial_rank(const knot_rrset_t *rr,
 		/* TODO: this classifier of authoritativity may not be perfect yet. */
 		rank |= KR_RANK_AUTH;
 	}
-	return (uint8_t)rank;
+	return rank;
 }
 
 static int pick_authority(knot_pkt_t *pkt, struct kr_request *req, bool to_wire)
